assets/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=0.1.2",
    "strands-agents>=1.4.0",
]

[dependency-groups]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.1.6",
]
